HomeWork 02-04 working
