+++ src/by/it/yarmolenka/MathCalc/CalcException.java
@@ -1,0 +1,18 @@
+package by.it.yarmolenka.MathCalc;
+
+public class CalcException extends Exception {
+    public CalcException() {
+    }
+
+    public CalcException(String message) {
+        super("ERROR:" + message);
+    }
+
+    public CalcException(String message, Throwable cause) {
+        super("ERROR:" + message, cause);
+    }
+
+    public CalcException(Throwable cause) {
+        super(cause);
+    }
+}
