--- src/by/it/yarmolenka/MathCalc_v1_severalCalcuations_singleBrackets_Exceptions/CalcException.java
@@ -1,18 +1,0 @@
-package by.it.yarmolenka.MathCalc_v1_severalCalcuations_singleBrackets_Exceptions;
-
-public class CalcException extends Exception {
-    public CalcException() {
-    }
-
-    public CalcException(String message) {
-        super("ERROR: " + message);
-    }
-
-    public CalcException(String message, Throwable cause) {
-        super("ERROR: " + message, cause);
-    }
-
-    public CalcException(Throwable cause) {
-        super(cause);
-    }
-}
