--- src/by/it/yarmolenka/MathCalc_v1_severalCalcuations_singleBrackets_Exceptions/Operations/Muls/VectorMulScalar.java
@@ -1,13 +1,0 @@
-package by.it.yarmolenka.MathCalc_v1_severalCalcuations_singleBrackets_Exceptions.Operations.Muls;
-
-import by.it.yarmolenka.MathCalc_v1_severalCalcuations_singleBrackets_Exceptions.Variables.*;
-
-import java.util.Arrays;
-
-public class VectorMulScalar {
-    public static Vector vectorScalar(Vector vec, Scalar sc) {
-        double[] res = Arrays.copyOf(vec.value, vec.value.length);
-        for (int i = 0; i < res.length; i++) res[i] *= sc.value;
-        return new Vector(res);
-    }
-}
